complete ch01

+++ h2.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 
 // power: べき乗計算関数
-int power(int a, int i) {
+double power(int a, int i) {
   double power = 1;
   int counter = 0;
   while (counter < i) {
@@ -17,7 +17,7 @@ int main(){
     double arctan1 = 0;
     double d;
     while(i < 10000){
-      arctan1 += (double)power(-1,i) / (i*2 + 1);
+      arctan1 += power(-1,i) / (i*2 + 1);
       if(i % 100 == 0){
         printf("pi = %f\n", arctan1 * 4);
       }
